Misc: Necessary code changes for generating data files for EVE Crucible 1.0 expansion.

+++ EVEMon.Common/Data/EveProperty.cs
@@ -44,7 +44,7 @@ namespace EVEMon.Common.Data
         /// <summary>
         /// Gets the property's ID.
         /// </summary>
-        public int ID { get; private set; }
+        public long ID { get; private set; }
 
         /// <summary>
         /// Gets the property name (the displayName in the CCP tables).
@@ -74,7 +74,7 @@ namespace EVEMon.Common.Data
         /// <summary>
         /// Gets the unitID for this property.
         /// </summary>
-        public int UnitID { get; private set; }
+        public long UnitID { get; private set; }
 
         /// <summary>
         /// When true, the higher the value, the better it is.
