+++ Tools/XmlGenerator/StaticData/InvMarketGroup.cs
@@ -5,16 +5,16 @@ namespace EVEMon.XmlGenerator.StaticData
     public sealed class InvMarketGroup : IHasID
     {
         [XmlElement("marketGroupID")]
-        public int ID { get; set; }
+        public long ID { get; set; }
 
         [XmlElement("marketGroupName")]
         public string Name { get; set; }
 
         [XmlElement("parentGroupID")]
-        public int? ParentID { get; set; }
+        public long? ParentID { get; set; }
 
         [XmlElement("iconID")]
-        public int? IconID { get; set; }
+        public long? IconID { get; set; }
 
         [XmlElement("description")]
         public string Description { get; set; }
